• Fixed issue when pulling information from SimPro    - Changed values from integers to double in Jobs and Quotes classes • Added ability to view if PO Number is missing

+++ DG Engineering/Framework/Global/SimPro/JSON Classes/Jobs.cs
@@ -79,7 +79,7 @@ namespace DG_Engineering.Framework.Global.SimPro
         public string Type { get; set; }
 
         [JsonProperty("TypeId")]
-        public int TypeId { get; set; }
+        public double TypeId { get; set; }
     }
 
     public class ProjectManager
@@ -94,7 +94,7 @@ namespace DG_Engineering.Framework.Global.SimPro
         public string Type { get; set; }
 
         [JsonProperty("TypeId")]
-        public int TypeId { get; set; }
+        public double TypeId { get; set; }
     }
 
     public class Status
@@ -125,16 +125,16 @@ namespace DG_Engineering.Framework.Global.SimPro
         public double Actual { get; set; }
 
         [JsonProperty("Committed")]
-        public int Committed { get; set; }
+        public double Committed { get; set; }
 
         [JsonProperty("Estimate")]
         public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class ConvertedFromQuote
@@ -167,16 +167,16 @@ namespace DG_Engineering.Framework.Global.SimPro
         public double Actual { get; set; }
 
         [JsonProperty("Committed")]
-        public int Committed { get; set; }
+        public double Committed { get; set; }
 
         [JsonProperty("Estimate")]
         public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class Labor
@@ -185,16 +185,16 @@ namespace DG_Engineering.Framework.Global.SimPro
         public double Actual { get; set; }
 
         [JsonProperty("Committed")]
-        public int Committed { get; set; }
+        public double Committed { get; set; }
 
         [JsonProperty("Estimate")]
         public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class LaborHours
@@ -203,67 +203,67 @@ namespace DG_Engineering.Framework.Global.SimPro
         public double Actual { get; set; }
 
         [JsonProperty("Committed")]
-        public int Committed { get; set; }
+        public double Committed { get; set; }
 
         [JsonProperty("Estimate")]
-        public int Estimate { get; set; }
+        public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class PlantAndEquipment
     {
         [JsonProperty("Actual")]
-        public int Actual { get; set; }
+        public double Actual { get; set; }
 
         [JsonProperty("Committed")]
-        public int Committed { get; set; }
+        public double Committed { get; set; }
 
         [JsonProperty("Estimate")]
-        public int Estimate { get; set; }
+        public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class PlantAndEquipmentHours
     {
         [JsonProperty("Actual")]
-        public int Actual { get; set; }
+        public double Actual { get; set; }
 
         [JsonProperty("Estimate")]
-        public int Estimate { get; set; }
+        public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class Overhead
     {
         [JsonProperty("Actual")]
-        public int Actual { get; set; }
+        public double Actual { get; set; }
 
         [JsonProperty("Committed")]
-        public int Committed { get; set; }
+        public double Committed { get; set; }
 
         [JsonProperty("Estimate")]
-        public int Estimate { get; set; }
+        public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class ResourcesCost
@@ -296,10 +296,10 @@ namespace DG_Engineering.Framework.Global.SimPro
         public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class ResourcesMarkup
@@ -317,16 +317,16 @@ namespace DG_Engineering.Framework.Global.SimPro
     public class Adjusted
     {
         [JsonProperty("Actual")]
-        public int Actual { get; set; }
+        public double Actual { get; set; }
 
         [JsonProperty("Estimate")]
-        public int Estimate { get; set; }
+        public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class GrossProfitLoss
@@ -338,10 +338,10 @@ namespace DG_Engineering.Framework.Global.SimPro
         public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class GrossMargin
@@ -353,10 +353,10 @@ namespace DG_Engineering.Framework.Global.SimPro
         public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class NettProfitLoss
@@ -368,10 +368,10 @@ namespace DG_Engineering.Framework.Global.SimPro
         public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class NettMargin
@@ -383,10 +383,10 @@ namespace DG_Engineering.Framework.Global.SimPro
         public double Estimate { get; set; }
 
         [JsonProperty("Revised")]
-        public int Revised { get; set; }
+        public double Revised { get; set; }
 
         [JsonProperty("Revized")]
-        public int Revized { get; set; }
+        public double Revized { get; set; }
     }
 
     public class Totals
@@ -407,16 +407,16 @@ namespace DG_Engineering.Framework.Global.SimPro
         public Adjusted Adjusted { get; set; }
 
         [JsonProperty("MembershipDiscount")]
-        public int MembershipDiscount { get; set; }
+        public double MembershipDiscount { get; set; }
 
         [JsonProperty("Discount")]
-        public int Discount { get; set; }
+        public double Discount { get; set; }
 
         [JsonProperty("STCs")]
-        public int STCs { get; set; }
+        public double STCs { get; set; }
 
         [JsonProperty("VEECs")]
-        public int VEECs { get; set; }
+        public double VEECs { get; set; }
 
         [JsonProperty("GrossProfitLoss")]
         public GrossProfitLoss GrossProfitLoss { get; set; }
@@ -473,10 +473,10 @@ namespace DG_Engineering.Framework.Global.SimPro
         public bool VEECsEligible { get; set; }
 
         [JsonProperty("STCValue")]
-        public int STCValue { get; set; }
+        public double STCValue { get; set; }
 
         [JsonProperty("VEECValue")]
-        public int VEECValue { get; set; }
+        public double VEECValue { get; set; }
     }
 
     public class Root
